refactor chat con chatbot para respetar arquitectura mvc
