+++ AdministradorChatBot/Controllers/HomeController.cs
@@ -23,32 +23,13 @@ public class HomeController(IChatbotService _chatbotService) : Controller
         return View();
     }
 
-    // CHAT CON EL CHATBOT -------- LOGICA DE HISTORIAL DEL CHAT
-    private const string ChatHistorySessionKey = "ChatHistory_";
-
-    private List<ChatMessage> GetChatHistory(int chatbotId)
-    {
-        var key = ChatHistorySessionKey + chatbotId;
-        var json = HttpContext.Session.GetString(key);
-        return json != null
-            ? JsonSerializer.Deserialize<List<ChatMessage>>(json) ?? new List<ChatMessage>()
-            : new List<ChatMessage>();
-    }
-
-    private void SaveChatHistory(int chatbotId, List<ChatMessage> history)
-    {
-        var key = ChatHistorySessionKey + chatbotId;
-        var json = JsonSerializer.Serialize(history);
-        HttpContext.Session.SetString(key, json);
-    }
-
     public async Task<IActionResult> Chat(int id)
     {
         var chatbot = await _chatbotService.GetChatbotWithKeywordsAndResponsesAsync(id);
         if (chatbot == null)
             return NotFound();
 
-        ViewBag.ChatHistory = GetChatHistory(id);
+        ViewBag.ChatHistory = _chatbotService.GetChatHistory(id, HttpContext.Session);
         return View(chatbot);
     }
 
@@ -57,32 +38,21 @@ public class HomeController(IChatbotService _chatbotService) : Controller
     {
         var chatbot = await _chatbotService.GetChatbotWithKeywordsAndResponsesAsync(chatbotId);
 
-        var history = GetChatHistory(chatbotId);
+        var history = _chatbotService.GetChatHistory(chatbotId, HttpContext.Session);
         history.Add(new ChatMessage { Sender = "Tú", Message = userMessage });
 
-        var keyword = chatbot?.ChatbotKeywords
-            .FirstOrDefault(k => string.Equals(k.Keyword, userMessage, StringComparison.OrdinalIgnoreCase));
-
-        string responseText;
-        if (keyword != null && keyword.ChatbotResponses.Any())
-        {
-            var random = new Random();
-            var responses = keyword.ChatbotResponses;
-            var randomResponse = responses[random.Next(responses.Count)];
-            responseText = randomResponse.Response;
-        }
-        else
-        {
-            responseText = "No tengo una respuesta para eso.";
-        }
+        var responseText = _chatbotService.GetChatbotResponse(chatbot, userMessage);
 
         history.Add(new ChatMessage { Sender = "Chatbot", Message = responseText });
-        SaveChatHistory(chatbotId, history);
+        _chatbotService.SaveChatHistory(chatbotId, history, HttpContext.Session);
 
         ViewBag.ChatHistory = history;
         return View("Chat", chatbot);
     }
-    // HASTA ACA LOGICA DEL CHAT E HISTORIAL
+
+
+
+
     [HttpPost]
     public IActionResult CreateChatbot(Chatbot chatbot)
     {
