+++ AdministradorChatBot/Interfaces/IChatbotService.cs
@@ -10,4 +10,8 @@ public interface IChatbotService
 
     void UpdateChatbot(Chatbot chatbot);
     Task DeleteChatbotAsync(Chatbot chatbot);
+
+    List<ChatMessage> GetChatHistory(int chatbotId, ISession session);
+    void SaveChatHistory(int chatbotId, List<ChatMessage> history, ISession session);
+    string GetChatbotResponse(Chatbot chatbot, string userMessage);
 }
